+++ crates/maelstrom-go-test/src/config.rs
@@ -1,0 +1,32 @@
+use maelstrom_macro::Config;
+
+#[derive(Config, Debug)]
+pub struct Config {
+    #[config(flatten)]
+    pub parent: maelstrom_test_runner::config::Config,
+
+    #[config(flatten, next_help_heading = "Go Testing Options")]
+    pub go_test_options: GoTestOptions,
+}
+
+#[derive(Config, Clone, Debug, Default)]
+pub struct GoTestOptions {
+    /// Controls the value of the `-vet` flag being passed to `go test`. See `go help test` for
+    /// details.
+    #[config(option, default = r#""go test's default""#, value_name = "VET_OPTIONS")]
+    pub vet: Option<String>,
+
+    /// Tells long-running tests to shorten their run time. See `go help testflag` for details.
+    #[config(flag)]
+    pub short: bool,
+
+    /// Shows the full file names in error messages. See `go help testflag` for details.
+    #[config(flag)]
+    pub fullpath: bool,
+}
+
+impl AsRef<maelstrom_test_runner::config::Config> for Config {
+    fn as_ref(&self) -> &maelstrom_test_runner::config::Config {
+        &self.parent
+    }
+}