+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `maelstrom-go-test` now edits out some lines from test failures that come from the `go test` test
   fixture code.
 - `maelstrom-go-test` now shows tests which call `t.Skip` as status `IGNORED`
+- `--vet` option added to `maelstrom-go-test`
+- `--short` flag added to `maelstrom-go-test`
+- `--fullpath` flag added to `maelstrom-go-test`
 
 #### Changed
 - `ui` test-runner configuration option value now matches the same casing as the CLI option